Move LogWindow to new files
