+++ src/menus/HelpMenus.cpp
@@ -6,12 +6,12 @@
 
 #include "../AboutDialog.h"
 #include "../AllThemeResources.h"
-#include "../AudacityLogger.h"
 #include "../AudioIOBase.h"
 #include "../CommonCommandFlags.h"
 #include "../CrashReport.h" // for HAS_CRASH_REPORT
 #include "../FileNames.h"
 #include "../HelpText.h"
+#include "../LogWindow.h"
 #include "../Menus.h"
 #include "../Prefs.h"
 #include "../Project.h"
